Controls checked.

+++ patika-06-01/src/main/java/dev/patika/patika0601/service/EmployeeService.java
@@ -1,5 +1,7 @@
 package dev.patika.patika0601.service;
 
+import dev.patika.patika0601.exceptions.EmployeeAlreadyExistsException;
+import dev.patika.patika0601.exceptions.EmployeeNotFoundException;
 import dev.patika.patika0601.model.Employee;
 import dev.patika.patika0601.repository.EmployeeRepository;
 import lombok.AllArgsConstructor;
@@ -17,12 +19,14 @@ public class EmployeeService {
     }
 
     public void addEmployee(Employee employee) {
-        // TO-DO check if employee exists???
+        if(this.repository.existsAllByEmail(employee.getEmail()))
+            throw new EmployeeAlreadyExistsException("Employee with E-mail: "+ employee.getEmail() + " is already exists!");
         repository.save(employee);
     }
 
     public void deleteEmployee(Long employeeId) {
-        // TO-DO check if employee exists with given ID???
+        if(!this.repository.existsAllById(employeeId))
+            throw new EmployeeNotFoundException("Employee with ID: " + employeeId + " could not foud!");
         repository.deleteById(employeeId);
 
     }
